Fix knowledge-base tests
