+++ app-modules/knowledge-base/tests/Tenant/KnowledgeBaseCategory/RequestFactories/CreateKnowledgeBaseCategoryRequestFactory.php
@@ -34,16 +34,17 @@
 </COPYRIGHT>
 */
 
-namespace AidingApp\KnowledgeBase\Tests\KnowledgeBaseQuality\RequestFactories;
+namespace AidingApp\KnowledgeBase\Tests\Tenant\KnowledgeBaseCategory\RequestFactories;
 
 use Worksome\RequestFactories\RequestFactory;
 
-class CreateKnowledgeBaseQualityRequestFactory extends RequestFactory
+class CreateKnowledgeBaseCategoryRequestFactory extends RequestFactory
 {
     public function definition(): array
     {
         return [
-            'name' => $this->faker->word(),
+            'name' => fake()->word(),
+            'slug' => fake()->slug(),
         ];
     }
 }
