+++ app-modules/knowledge-base/tests/Tenant/KnowledgeBaseCategory/RequestFactories/EditKnowledgeBaseCategoryRequestFactory.php
@@ -34,7 +34,7 @@
 </COPYRIGHT>
 */
 
-namespace AidingApp\KnowledgeBase\Tests\KnowledgeBaseCategory\RequestFactories;
+namespace AidingApp\KnowledgeBase\Tests\Tenant\KnowledgeBaseCategory\RequestFactories;
 
 use Worksome\RequestFactories\RequestFactory;
