+++ app-modules/knowledge-base/tests/Tenant/KnowledgeBaseItem/EditKnowledgeBaseItemTest.php
