+++ app-modules/knowledge-base/tests/Tenant/KnowledgeBaseQuality/ViewKnowledgeBaseQualityTest.php
